requirements.txt
streamlit
openpyxl
python-barcode
Pillow